Solve day 12
Sadly I got a bit spoiled. I had thought of this as a heuristic that
might work, but thought something more complicated might be needed as
well. Then I saw several people saying it was too simple and that they
felt dirty, so I tried this. Not the most fun last day.